# webapitc2005b/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Web API Express",
        "description": "My WebApi using Node",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3001",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/getUsuario/{id}": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Encuentra un usuario por id",
                "description": "Regresa el usuario con el id solicitado",
                "operationId": "getUsuario",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Id del usuario a consultar",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Usuario"
                        }
                    }
                }
            }
        },
        "/api/addUsuario": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Agrega un nuevo usuario",
                "description": "Agrega un usuario a la base de datos",
                "operationId": "addUsuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nuevo jugador",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Usuario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Usuario"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/createUser": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Agrega un nuevo usuario",
                "description": "Agrega un usuario a la base de datos",
                "operationId": "createUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nuevo jugador",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Usuario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Usuario"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/updateUsuario/{id}": {
            "put": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Actualiza un usuario",
                "description": "Actualiza un usuario a la base de datos",
                "operationId": "updateUsuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Id del usuario a actualizar",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos nuevos usuario",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Usuario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Usuario"
                        }
                    },
                    "400": {
                        "description": "Jugador Inválido"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            }
        },
...
        "/api/getUserType/{user_type_id}": {
            "get": {
                "tags": [
                    "UserType"
                ],
                "summary": "Encuentra un tipo de usuario por id",
                "description": "Regresa el tipo de usuario con el id solicitado",
                "operationId": "getUserType",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_type_id",
                        "description": "Id del tipo de usuario a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/UserType"
                        }
                    }
                }
            }
        },
        "/api/addUserType": {
            "post": {
                "tags": [
                    "UserType"
                ],
                "summary": "Agrega un nuevo tipo de usuario",
                "description": "Agrega un tipo de usuario a la base de datos",
                "operationId": "addUserType",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nuevo grupo",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/UserType"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/updateUserType/{user_type_id}": {
            "put": {
                "tags": [
                    "UserType"
                ],
                "summary": "Actualiza un tipo de usuario",
                "description": "Actualiza un tipo de usuario en la base de datos",
                "operationId": "updateUserType",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_type_id",
                        "description": "Id del tipo de usuario a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos nuevo tipo de usuario",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/UserType"
                        }
                    },
                    "400": {
                        "description": "Tipo de usuario Inválido"
                    },
                    "404": {
                        "description": "Tipo de usuario no encontrado"
                    }
                }
            }
        },
        "/api/deleteUserType/{user_type_id}": {
            "delete": {
                "tags": [
                    "UserType"
                ],
                "summary": "Borra un tipo de usuario",
                "description": "Borra un tipo de usuario de la base de datos",
                "operationId": "deleteUserType",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_type_id",
                        "description": "Id del tipo de usuario a borrar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/UserType"
                        }
                    },
                    "400": {
                        "description": "Tipo de Usuario Inválido"
                    },
                    "404": {
                        "description": "Tipo de Usuario no encontrado"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Usuario": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "id_grupo": {
                    "type": "integer",
                    "format": "int64"
                },
                "tipo": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "estatus": {
                    "type": "string"
                }
            }
        },
...
        "UserType": {
            "type": "object",
            "properties": {
                "user_type_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "user_type_desc": {
                    "type": "string"
                }
            }
        },
...
    }
}